restructure _utils and utils submodules
