+++ hydromt/_utils/rgetattr.py
@@ -1,7 +1,9 @@
 from functools import reduce
 
+__all__ = ["_rgetattr"]
 
-def rgetattr(obj, attr, *args):
+
+def _rgetattr(obj, attr, *args):
     """Recursive get attribute from object."""
 
     def _getattr(obj, attr):
