+++ hydromt/_utils/uris.py
@@ -3,13 +3,15 @@ from urllib.parse import urlparse
 
 from hydromt._typing import StrPath
 
+__all__ = ["_strip_scheme", "_is_valid_url"]
 
-def strip_scheme(uri: str) -> str:
+
+def _strip_scheme(uri: str) -> str:
     """Strip scheme from uri."""
     return re.sub(r"^\w+://", "", uri)
 
 
-def is_valid_url(uri: StrPath) -> bool:
+def _is_valid_url(uri: StrPath) -> bool:
     """Check if uri is valid."""
     try:
         result = urlparse(str(uri))
