+++ hydromt/data_catalog/adapters/__init__.py
@@ -1,8 +1,7 @@
 """Data Adapters are generic for its HydroMT type and perform transformations."""
 
 # TODO: correct imports after deprecation of old adapters
-from .caching import cache_vrt_tiles
-from .data_adapter import PREPROCESSORS, DataAdapter
+from .data_adapter import DataAdapter
 from .dataframe import DataFrameAdapter
 from .dataset import DatasetAdapter
 from .geodataframe import GeoDataFrameAdapter
@@ -16,6 +15,4 @@ __all__ = [
     "GeoDatasetAdapter",
     "RasterDatasetAdapter",
     "DatasetAdapter",
-    "cache_vrt_tiles",
-    "PREPROCESSORS",
 ]
