+++ hydromt/data_catalog/drivers/geodataset/vector_driver.py
@@ -9,7 +9,7 @@ from xarray import DataArray, Dataset
 from hydromt._typing import CRS, SourceMetadata
 from hydromt._typing.error import NoDataStrategy
 from hydromt._typing.type_def import Geom, Predicate, TimeRange
-from hydromt._utils.unused_kwargs import warn_on_unused_kwargs
+from hydromt._utils.unused_kwargs import _warn_on_unused_kwargs
 from hydromt.data_catalog.drivers.geodataset.geodataset_driver import GeoDatasetDriver
 from hydromt.data_catalog.drivers.preprocessing import PREPROCESSORS
 from hydromt.io import open_geodataset
@@ -41,7 +41,7 @@ class GeoDatasetVectorDriver(GeoDatasetDriver):
 
         Args:
         """
-        warn_on_unused_kwargs(
+        _warn_on_unused_kwargs(
             self.__class__.__name__,
             {
                 "predicate": predicate,
