+++ hydromt/data_catalog/drivers/geodataset/xarray_driver.py
@@ -16,7 +16,7 @@ from hydromt._typing import (
 )
 from hydromt._typing.error import NoDataStrategy
 from hydromt._typing.type_def import Predicate
-from hydromt._utils.unused_kwargs import warn_on_unused_kwargs
+from hydromt._utils.unused_kwargs import _warn_on_unused_kwargs
 from hydromt.data_catalog.drivers.geodataset.geodataset_driver import GeoDatasetDriver
 from hydromt.data_catalog.drivers.preprocessing import PREPROCESSORS
 
@@ -48,7 +48,7 @@ class GeoDatasetXarrayDriver(GeoDatasetDriver):
 
         Args:
         """
-        warn_on_unused_kwargs(
+        _warn_on_unused_kwargs(
             self.__class__.__name__,
             {
                 "mask": mask,
