+++ hydromt/data_catalog/drivers/raster/raster_xarray_driver.py
@@ -17,7 +17,7 @@ from hydromt._typing import (
     ZoomLevel,
 )
 from hydromt._typing.error import NoDataStrategy
-from hydromt._utils.unused_kwargs import warn_on_unused_kwargs
+from hydromt._utils.unused_kwargs import _warn_on_unused_kwargs
 from hydromt.data_catalog.drivers.preprocessing import PREPROCESSORS
 from hydromt.data_catalog.drivers.raster.raster_dataset_driver import (
     RasterDatasetDriver,
@@ -50,7 +50,7 @@ class RasterDatasetXarrayDriver(RasterDatasetDriver):
 
         Args:
         """
-        warn_on_unused_kwargs(
+        _warn_on_unused_kwargs(
             self.__class__.__name__,
             {
                 "mask": mask,
