+++ hydromt/utils/__init__.py
@@ -1,5 +1,6 @@
 """Public helper functions that (pulgin) developers can use."""
 
 from hydromt.utils.deep_merge import deep_merge
+from hydromt.utils.log import setuplog
 
-__all__ = ["deep_merge"]
+__all__ = ["deep_merge", "setuplog"]
