+++ tests/_utils/test_rgetattr.py
@@ -2,18 +2,18 @@ import os
 
 import pytest
 
-from hydromt._utils.rgetattr import rgetattr
+from hydromt._utils.rgetattr import _rgetattr
 
 
 def test_rgetattr_os_sep():
-    assert rgetattr(os, "path.sep") == os.path.sep
+    assert _rgetattr(os, "path.sep") == os.path.sep
 
 
 def test_rgetattr_os_fake_default():
-    assert rgetattr(os, "path.sepx", None) is None
-    assert rgetattr(os, "path.sepx", "--") == "--"
+    assert _rgetattr(os, "path.sepx", None) is None
+    assert _rgetattr(os, "path.sepx", "--") == "--"
 
 
 def test_rgetattr_os_fake_no_default_fails():
     with pytest.raises(AttributeError):
-        rgetattr(os, "path.sepx")
+        _rgetattr(os, "path.sepx")
